+++ core/auth/build.gradle.kts
@@ -28,5 +28,4 @@ dependencies {
     implementation(libs.androidx.security.crypto)
     implementation(project(":core:model"))
 
-    implementation(project(":feature:profile-api"))
 }
