+++ core/auth/src/main/kotlin/su/afk/kemonos/auth/ClearAllAuthUseCase.kt
@@ -1,0 +1,12 @@
+package su.afk.kemonos.auth
+
+import su.afk.kemonos.auth.domain.repository.AuthRepository
+import javax.inject.Inject
+
+class ClearAllAuthUseCase @Inject constructor(
+    private val authRepository: AuthRepository,
+) {
+    suspend operator fun invoke() {
+        authRepository.clearAll()
+    }
+}
