+++ core/auth/src/main/kotlin/su/afk/kemonos/auth/di/AuthPrefsModule.kt
@@ -16,7 +16,7 @@ import javax.inject.Singleton
 
 @Module
 @InstallIn(SingletonComponent::class)
-internal object AuthPrefsModule {
+internal object AuthDataStoreModule {
 
     @Provides
     @Singleton
