+++ core/auth/src/main/kotlin/su/afk/kemonos/auth/domain/repository/AuthSessionProvider.kt
@@ -1,0 +1,7 @@
+package su.afk.kemonos.auth.domain.repository
+
+import su.afk.kemonos.domain.SelectedSite
+
+interface AuthSessionProvider {
+    suspend fun getSession(site: SelectedSite): String?
+}
