--- core/auth/src/main/kotlin/su/afk/kemonos/auth/news/CtyptoManager.kt
@@ -1,11 +1,0 @@
-package su.afk.kemonos.auth.news
-
-import java.security.KeyStore
-
-class CtyptoManager {
-    private val keyStore = KeyStore.getInstance(KeyStore.getDefaultType()).apply {
-        load(null)
-    }
-
-
-}
