+++ core/model/src/main/kotlin/su/afk/kemonos/domain/models/AuthUser.kt
@@ -1,0 +1,11 @@
+package su.afk.kemonos.domain.models
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class AuthUser(
+    val id: Int,
+    val username: String,
+    val createdAt: String,
+    val role: String,
+)
