+++ gradle/libs.versions.toml
@@ -25,7 +25,7 @@ credentials = "1.5.0"
 # --- Compose
 # Для BOM не задавать вручную версии для compose libs
 # foundation/material3/icons/etс должно подтягиваться BOM
-composeBom = "2026.03.00"
+composeBom = "2026.03.01"
 
 # --- DI
 hilt = "2.58"
